refactor(console): reorg connector-related files
