+++ packages/console/src/hooks/use-connector-form-config-parser.tsx
@@ -2,8 +2,8 @@ import type { ConnectorResponse } from '@logto/schemas';
 import { toast } from 'react-hot-toast';
 import { useTranslation } from 'react-i18next';
 
-import { parseFormConfig } from '@/pages/Connectors/components/ConnectorForm/utils';
-import type { ConnectorFormType } from '@/pages/Connectors/types';
+import type { ConnectorFormType } from '@/types/connector';
+import { parseFormConfig } from '@/utils/connector-form';
 import { safeParseJson } from '@/utils/json';
 
 const useJsonStringConfigParser = () => {
