+++ packages/console/src/pages/ConnectorDetails/ConnectorTabs/index.module.scss
