+++ packages/console/src/pages/ConnectorDetails/ConnectorTabs/index.tsx
