+++ packages/console/src/pages/ConnectorDetails/ConnectorTypeName/index.tsx
