+++ packages/console/src/pages/Connectors/ConnectorName/index.module.scss
