+++ packages/console/src/pages/Connectors/ConnectorName/index.tsx
