+++ packages/console/src/pages/Connectors/CreateForm/PlatformSelector/index.module.scss
