+++ packages/console/src/pages/Connectors/CreateForm/PlatformSelector/index.tsx
@@ -6,7 +6,7 @@ import { connectorPlatformLabel } from '@/consts/connectors';
 import RadioGroup, { Radio } from '@/ds-components/RadioGroup';
 import type { ConnectorGroup } from '@/types/connector';
 
-import * as styles from './PlatformSelector.module.scss';
+import * as styles from './index.module.scss';
 
 type Props = {
   connectorGroup: ConnectorGroup<ConnectorFactoryResponse & { added: boolean }>;
