+++ packages/console/src/pages/Connectors/SignInExperienceSetupNotice/index.tsx
