+++ packages/console/src/types/connector.ts
@@ -7,3 +7,17 @@ export type ConnectorGroup<T = ConnectorResponse> = Pick<
   id: string;
   connectors: T[];
 };
+
+export enum SyncProfileMode {
+  OnlyAtRegister = 'OnlyAtRegister',
+  EachSignIn = 'EachSignIn',
+}
+
+export type ConnectorFormType = {
+  config: string;
+  name: string;
+  logo: string;
+  logoDark: string;
+  target: string;
+  syncProfile: SyncProfileMode;
+} & Record<string, unknown>; // Extend custom connector config form
