+++ packages/console/src/utils/connector-form.ts
@@ -1,9 +1,8 @@
 import type { ConnectorConfigFormItem } from '@logto/connector-kit';
 import { ConnectorConfigFormItemType } from '@logto/connector-kit';
 
+import type { ConnectorFormType } from '@/types/connector';
 import { safeParseJson } from '@/utils/json';
-
-import type { ConnectorFormType } from '../../types';
 
 export const initFormData = (
   formItems: ConnectorConfigFormItem[],
